feat: working on profile functions
