+++ client/src/components/testUserProfile/testUserProfile.js
@@ -1,70 +1,9 @@
-import portadaUno from "../../assets/images/portada-1.png";
-import portadaDos from "../../assets/images/portada-2.png";
-import portadaTres from "../../assets/images/portada-3.png";
-import portadaCuatro from "../../assets/images/portada-4.png";
-import { BsFillCaretRightFill } from "react-icons/bs";
-import { FaRegHeart } from "react-icons/fa";
 import Input from "../Input";
 import Button from "../Buttons/index";
-import { getCurrentUser, reauthenticate } from "../../services/auth";
-import { useState, useEffect } from "react";
-import { useHistory } from "react-router";
-import { useDispatch, useSelector } from "react-redux";
 
 import "./style/userprofile.scss";
-import {
-  updateUserProfileInfo,
-  updateUserProfilePassword,
-} from "../../redux/user/action";
 
 function UserProfile() {
-  const authUserState = useSelector((state) => state.auth.user);
-  const history = useHistory();
-  const { loading, accessToken, signOutSuccess } = useSelector(
-    (state) => state.auth
-  );
-
-  useEffect(() => {
-    if (!loading && !accessToken && signOutSuccess) {
-      setProfile({ email: "", firstName: "", lastName: "" });
-      history.push("/login");
-    }
-  }, [loading, accessToken, signOutSuccess, history]);
-
-  useEffect(() => {
-    setProfile({
-      email: authUserState.email,
-      firstName: authUserState.firstName,
-      lastName: authUserState.lastName,
-    });
-  }, []);
-  const dispatch = useDispatch();
-
-  const [profile, setProfile] = useState({
-    email: "",
-    firstName: "",
-    lastName: "",
-  });
-
-  const [password, setPassword] = useState("");
-  function handleUserInfoSubmit(e) {
-    e.preventDefault();
-
-    const userId = getCurrentUser().uid;
-    dispatch(updateUserProfileInfo(userId, profile));
-  }
-  function handlePasswordSubmit(e) {
-    e.preventDefault();
-    dispatch(updateUserProfilePassword(password));
-  }
-  function handleProfileChange(e) {
-    setProfile({ ...profile, [e.target.name]: e.target.value });
-    console.log(e.target.value, "CHANGE");
-  }
-  function handlePasswordChange(e) {
-    setPassword(e.target.value);
-  }
-
   return (
     <>
       <div className="profile-container">
